templatemanager: Fix empty dialog when saving templates.
We used to call just mpTabControl->Clear() in setSaveMode(); but that worked
only with .src dialogs, as the tab pages could have existed even without
TabControl containing them.  This is not possible with .ui definitions any
more (and rightly so!), so show just one tab here for now, until we do a
bigger Template manager rework.

Change-Id: I400dbcc326e883a1c1f526e94974162c3641f8ae