related: fdo#74230: Area tab page: prevent default gradient/hatch/bitmap
The Area dialog for shapes unfortunately makes it possible to set
gradients/hatching/bitmap that uses the default items in the
SfxItemPool.  These items however cannot be stored to ODF files, since
these are represented as elements, not as attributes on a
style:default-style; what we get for defaults is a somewhat silly
draw:fill="hatch" without an draw:fill-hatch-name.
So prevent the dialog from creating them by forcing a selection of an
entry in the list.

Change-Id: I67cc6dbbf7b491f06d094d4de1e9c3ffe79b01f5