fdo#52547 SOT: Prefer embedding image data to embedding linked image.
When copying an image from e.g. Firefox or Chromium the clipboard
will contain both the image itself, and the url of the image --
presumably it makes more sense to store the image data itself if
it is available -- the order of the items in the *_FREE_AREA_Copy
defines the order of preference for which data in the clipboard
should be used.

Change-Id: Id8b862a9f2fce3a34adbb16509e9b7372500a239
Reviewed-on: https://gerrit.libreoffice.org/7928
Tested-by: Caolán McNamara <caolanm@redhat.com>
Reviewed-by: Caolán McNamara <caolanm@redhat.com>