New frame attribute: TextVerticalAdjust <-> RES_TEXT_VERT_ADJUST
By now this property is added for MS Word interoperability.
With it we can make DOCX import/export working.
I use this solution instead of InteropGrabBag because vertical
adjustment would be a good thing to have for text frames too.
So later it can be the first step of an enhanchement.

Change-Id: I9dcfdc8ae56ce443f0c286c51136e1d006951938