startcenter: don't generate thumbnails in headless mode
Commit f9c0624c55b2d3ecdd46754dfd6db586ef14fa6c (startcenter: fdo#72469:
Thumbnails also for other file types than ODF, 2014-02-03) turned on
thumbnail generation for non-ODF documents as well. This is not that
interesting in case --headless is used, however it slows down unit tests
quite a lot, resulting in about doubled toplevel noop make time.

Numbers on my machine for toplevel 'make' (in case there is no code to
rebuild):
- before: real    5m56.857s
- after:  real    2m39.432s

Change-Id: I0edc13c29a7713f8faf8158ba240151977cdfa4b