fdo#74931 Prevent UpdateSettings from dereferencing null pointer
When calling on GtkSalFrame::GetGraphics(), if all graphics are in
use then it will return NULL. We don't currently check for this,
but we still try to use the graphics object even if none are
available.

I have added a warning when this occurs and just return immediately,
as there's nothing more we can do.

Change-Id: I24dfbb6f37f41527078a60a7577cb3b20de3fcca
Reviewed-on: https://gerrit.libreoffice.org/8031
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>