drawingML import: fix inheritance of character height
The problem was that in case a shape had multiple (e.g. two) paragraphs,
and in case the first paragraph had an explicit character height, but
not the second, then the cursor carried over the explicit character
height to the second paragraph, but it shouldn't, as that leads to
incorrect character height in the second paragraph.

Fix this by remembering the default character height and using that in
case nothing is set explicitly.

Change-Id: I66e06d5cf192739fb254f7280c74617171d9ee6a