fdo#68927: sw: fix painting of SVG page background
There was a fix for a wrong constant resulting in too large image size in
5f9c7121e85dbd31b8a35ec53f81bb5f37ecaed1, and it turns out that painting
of SVG backgrounds was relying on the wasteful rasterization:  the SVG
is rasterized only once (at the initial zoom level) and then cached, and
now the correct image size for 100% is simply scaled up when zooming in.

Fix that by painting SVGs with the drawing layer primitives instead,
which appears to give better results.

Change-Id: I4be16856fd080290526d4963d8c512beefa85363