TestOutputDevice can't set meOutDevType
One of the StarMath tests was very broken - the issue is in formula tester. It
basically sets up its own derived TestOutputDevice. Because it doesn't set
meOutDevType (and can't as that's a private member function - I note this whole
design is *broken* but I digress) a few things can't get access to a
SalGraphics instance.  This is fixed by removing the class and changing it to a
VirtualDevice - the test runs under headless mode anyway

Change-Id: I3d6eba793386e6db3f053aa4e6aab9a4d1c63fc3
Reviewed-on: https://gerrit.libreoffice.org/8395
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>