unotools: improve date parser to support timezones
ISO8601 defines timezones with formats like 'Z', '+01', '-08:30' or
'+0830' at the end of the datetime string, but our parser was failing
when the string contained any timezone information.

I have modified the parser so it doesn't fail in case of the string
contains a timezone. Moreover, I check that the timezone definition is
correct according to the standard [1] and I extract it to the string
tokTz.

Unfortunately UNO Time class doesn't contain any field to store the
timezone, only the boolean field IsUTC, which I fill accordingly.

TODO: add a timezone field to UNO and fill it from the parser.

[1] https://en.wikipedia.org/wiki/ISO_8601#Time_zone_designators

Change-Id: I9d33b19efbc6d41a74b02ece2dfa12fa12fac5eb