writerfilter: introduce ooxml::OOXMLUniversalMeasureValue
In transitional DOCX, ST_UniversalMeasure is in practice a simple integer,
which means twips. But in case of strict, ST_UniversalMeasure is in points
in practice -- which is perfectly valid, but we didn't handle it so far.

Add a separate Value class that is used only for handling
ST_UniversalMeasure, then there we can handle the various additional
suffixes.

Change-Id: Iebb1ee859076595594d1455a1f826841dae77a0b