Avoid "delayed" rendering in the tiled rendering case
Obviously, when using tiled rendering, all of the tile should be
rendered at once during the call to lo_touch_draw_tile(). Nothing
should be delayed to happen later (when the PaintTimer fires). The
main culprit causing such "delayed" rendering, at least in a certain
test document, seems to be SmartArt objects.

It is not 100% clear how to determine whether tiled rendering is being
used, though. For now, let's assume it is always the case on
non-desktop (i.e. iOS and Android, even if for Android we dont
actually do it). For desktop platforms, including liblibreoffice, will
have to figure out later how we want to do it, assume never for now.

Change-Id: I0f7e7172e7a699a0c6752ee465776f9da5c3098e