fdo#74702 Moved ImplReleaseGraphics into correct classes
Made OutputDevice::ImplReleaseGraphics a pure virtual function, then
implemented function in Printer, Window and VirtualDevice. The reason
was that OutputDevice was checking to see if it was a Printer, Window
or VirtualDevice that was calling on it in an if statement, very
uncool :-) Now I let the classes themselves do the work.

There is some common functionality, which is to release the fonts. I
have put this into a protected OutputDevice function, ImplReleaseFonts.

Change-Id: Id41db2119d4022ea2fc7855158ca9f610af3c85c
Reviewed-on: https://gerrit.libreoffice.org/8548
Reviewed-by: Norbert Thiebaud <nthiebaud@gmail.com>
Tested-by: Norbert Thiebaud <nthiebaud@gmail.com>