ooxml: preserve several simple types of sdt
This patch is to preserve the sdt blocks for several simple sdt types:
equation, picture, citation and group.

This patch reuses the infrastructure created at commit
7d09ac9a2463d9e5f04b2e5ab811dcfdfd6f8d9d to grab-bag the properties of
the sdt block and write them back on export, with small modifications
in particular related with the fact that these tags don't have
children.

The type of the property tags equation, picture, citation and group
was modified in the ooxml model file to be able to process them. They
were previously empty.

A unit test was added to check the preservation of these fields.
Another had to be modified.

Change-Id: Ica1a340f589f5d6c85c1a35d2a3a4bf930eb2d8a