Add simple document selector to TiledLibreOffice
Make it easier to handle several test docs. Until now you had to
change the hardcoded document in the Xcode project and in the lo.mm
source file and re-build. Now it is enough to upload a new test doc to
the device using iTunes. If no test docs are present, use the good old
bundled test1.odt.

Change-Id: I3cbb9f74c17332ffc6ac90dd1e226fac005c3387